+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - **http client:** retain middleware logic when using `FluentHttpClient.RawHttpClient`
 - **http client:** add overload `Send(HttpRequestMessage request)` for better interoperability with the native when already having `HttpRequestMessage`
 - **headers:** `FluentHttpHeaders` now implements `IFluentHttpHeaderBuilder`, any extension methods added to `IFluentHttpHeaderBuilder` will be available for `FluentHttpHeaders` also, along with `FluentHttpClientBuilder` and `FluentHttpRequestBuilder`
+- **middleware context:** `FluentHttpMiddlewareContext` now includes formatters
 
 ## [3.2.1](https://github.com/sketch7/FluentlyHttpClient/compare/3.2.0...3.2.1) (2019-05-21)